+++ .gitignore
@@ -12,10 +12,6 @@ report.*.json
 # testing
 /coverage
 
-# next.js
-/.next/
-/out/
-
 # production
 /build
