+++ .prettierignore
@@ -2,7 +2,6 @@ coverage
 docs
 build
 publish
-storybook
 ._npm
 cx.json
 team-data
