+++ setupTests.js
@@ -1,2 +1,2 @@
-import '@testing-library/jest-dom/extend-expect'
-import '@testing-library/react-hooks/dont-cleanup-after-each.js'
+require ('@testing-library/jest-dom/extend-expect')
+require ('@testing-library/react-hooks/dont-cleanup-after-each.js')
